+++ src/spawn.spec.ts
@@ -13,7 +13,7 @@ describe("spawn", () => {
 
     it("will run the specified command", (done) => {
         const testFilePath = path.join(tmpDir.absPath(), "foo.txt");
-        spawn("touch", ["foo.txt"], tmpDir.absPath())
+        spawn("touch", ["foo.txt"], tmpDir.absPath()).closePromise
         .then(() => {
             const stats = fs.statSync(testFilePath);
             expect(stats.isFile()).toBeTruthy();
@@ -23,12 +23,27 @@ describe("spawn", () => {
 
 
     it("will resolve with the stdout text", (done) => {
-        spawn("touch", ["foo.txt"], tmpDir.absPath())
+        spawn("touch", ["foo.txt"], tmpDir.absPath()).closePromise
         .then(() => {
-            return spawn("ls", [], tmpDir.absPath());
+            return spawn("ls", [], tmpDir.absPath()).closePromise;
         })
         .then((output) => {
             expect(output).toContain("foo.txt");
+            done();
+        });
+    });
+
+
+    it("provides access to the underlying child process", (done) => {
+        const spawnResult = spawn("sleep", ["10"], tmpDir.absPath());
+        spawnResult.childProcess.kill();
+        spawnResult.closePromise
+        .then(() => {
+            fail("closePromise should reject when the child process is killed.");
+        })
+        .catch((reason) => {
+            expect(reason.exitCode).toEqual(null);
+            expect(reason.stderr).toEqual("");
             done();
         });
     });
@@ -97,7 +112,7 @@ describe("spawn", () => {
 
     it("provides the exit code and stderr when the command fails", (done) => {
         const nonExistantFilePath = path.join(tmpDir.absPath(), "xyzzy.txt");
-        spawn("cat", [nonExistantFilePath], ".")
+        spawn("cat", [nonExistantFilePath], ".").closePromise
         .catch((err) => {
             expect(err).toBeTruthy();
             expect(err.exitCode).not.toEqual(0);